Create bg icon generator function
