+++ src/components/index.js
@@ -1,5 +1,6 @@
 export { default as Menu } from './Menu/Menu';
-export { default as Background, generateBackground } from './ProjectBackground/ProjectBackground';
+export { default as Background } from './ProjectBackground/ProjectBackground';
+export { generateBackgroundIcons } from './ProjectBackground/backgroundGenerator';
 export { default as Welcome } from './Welcome/Welcome';
 export { default as About } from './About/About';
 export { default as Skills } from './Skills/Skills';
